refactor(services): extract the web loop context
Moves 14 methods plus the loop scheduler and its nine bindings into a
state module, a scheduler module and a client over a new interface file.
The scheduler keeps its timer semantics: it reads the phase delay and
the timer map through accessors rather than through shared bindings.

resetWebLoopsForTest is the one seam split across two owners, because it
also clears messagesBySession, which the composition root still owns; the
loop module exposes its two clear steps so the original ordering holds.

web-agent-client.ts 2373 -> 1877, agent-service.ts 384 -> 364.
